Move component specs into components folder
