+++ client/spec/component_spec/item.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import Item from '../src/components/item';
+import Item from '../../src/components/item';
 
 Enzyme.configure({adapter: new Adapter()});
