+++ client/spec/component_spec/itemList.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Enzyme from 'enzyme';
 import toJSON from 'enzyme-to-json';
 import Adapter from 'enzyme-adapter-react-16';
-import ItemList from '../src/components/itemList';
+import ItemList from '../../src/components/itemList';
 
 Enzyme.configure({adapter: new Adapter()});
