+++ client/spec/component_spec/itemSubmitForm.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import ItemSubmitForm from '../src/components/itemSubmitForm';
+import ItemSubmitForm from '../../src/components/itemSubmitForm';
 
 Enzyme.configure({adapter: new Adapter()});
