+++ client/spec/component_spec/userSignUpForm.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import UserSignUpForm from '../src/components/userSignUpForm';
+import UserSignUpForm from '../../src/components/userSignUpForm';
 
 Enzyme.configure({adapter: new Adapter()});
